Make usage of SequenceAligner static, impl. tests
The sequence aligner should be able to be used to align multiple pairs
of sequences and there is no need to save state for the alignments so
its methods are now static.

TestSequenceAligner has been implemented. Tests check that proper
exceptions are thrown if invalid sequences are passed in and check that
certain alignment cases work out as expected.