Add feature: Compare Proteins on aligned residues
Add feature to Metric class to construct itself from the aligned
residues in the two proteins passed in.

In support, add queries to get residue IDs and CA Distance Matrices based
on masks to specify which residues to include. Also add method to
SequenceAligner to produce these masks from alignment strings. This
method takes as input alignment strings produced by the align methods.
In the future, a method can be added that takes the proteins and
produced the masks directly for applications that don't need the
alignments.

Minor change in PDBFileIO. Remove a print statement.