Update JProtMetrics Usage Information
Clarify the usage message and move it into its own file as a resource.
All executables will now handle usage information this way. Usage info
goes in a file and the executable reads it in to print it. This makes it
easier to edit the usage info and handle formatting.

JProtMetrics now determines whether to read in CA Distance Matrices or
PDB files based on the extension of the files.