Refactor PDBFileIO
Moved the logic of building the residues, chains, and the protein out of
the method to read in the PDBFile and into its own specific method. The
readInPDBFile method reads in the records and saves them in lists. These
lists are then used by the helper buildProtein method to construct the
Protein. This moves and allows for better management of the complexity
of the process. Having PDBFileIO read in and handle other types of
records in pdb files will now be easier. For example, the next task is
to read in the SSBOND records and add those bonds to the structures.