Move Graph representation into Protein, refac IO
Working on adding disulfide bridges to the protein model. Since
disulfide bridges can be between chains, that information is at the
protein level and since the graph represents all covalent bonds within
a molecule that graph should be at the protein level since chains can be
covalently linked.

Along the way, testing indicated that the protein graph isn't being
built correctly. Properly debugging this requires junit tests for
individual residues. These tests should have been here, but that's the
current task. To test individual residues, it would be ideal to be able
to read in any of the default residues (this feature can be used later
to build proteins from sequence data alone). This prompted a rethink and
refactoring of the PDB File input process.

PDBFileIO should only be responsible for reading in and writing out pdb
records, not for constructing proteins as well. Every protein should
construct it's own instance of PDBFileIO that it can use to get the
relevant records to construct itself. It can also use PDBFileIO to write
itself out to file.