Make PDBFileIO a component of Protein
Now protein's constructor takes an inputStream pointing to a pdb as
input, constructs its own PDBFileIO and uses that IO object to construct
itself. This required changes in the tester and the executable
JProtMetrics to accommodate the new usage. Now, when a protein needs to
be isntantiated, the client does not need to build the PDBFileIO object.
The process is simplified.