Add Classes for Reading and Manipulating PMFs
Add package energy with PotentialOfMeanForce. This class holds all
various information that represents a PMF. It includes the ability to
get and set energy functions for atom type pairs.

Add class PMFFileIO to io package with two public methods to read in and
write out our PMFs in the encad PMF file format.