Move Protein Building Logic into PDBFileIO
If we view the pdb as a kind of view for a protein, then the model (the
protein) shouldn't need to know the details of the view (the output file
manager), so all details for the building of a protein from the PDB go
into PDBFileIO. A PDBFileIO object can be used to read in and return
a protein. A single instance of PDBFileIO can only every read in
1 protein since it is responsible for managing all the extra metadata
that is included in a PDB file but not relevant for constructing a model
of a protein.

Every protein will keep a reference to a PDBFileIO object. If the
protein was read in from a file, this is the PDBFileIO instance that was
used to read it in. If the protein was built ab initio, it is a default
PDBFileIO object that can be used to write the protein out to file.