Change Atom and Residue Constructors
Change Atom constructors to take in a String representing the element of
this atom. As of this commit, that can be H, C, N, O, or S. In the
future this list may be expanded.

Modify the Residue constructor that instantiates a default residue to
actually do so. It uses the PDBFileIO static method
getDefaultResidueAtoms(resName) to get the atoms for the default residue
so that it can now be constructed.