Complete TestResidue
TestResidue tests the construction of all the default residues and
protonation states. It checks that each residue has all the correct
atoms, it checks that the setAsCarboxylTerminus method works, and it
checks that each residue has the correct number of atoms and bonds both
before and after both enabling and disabling hydrogens (and ensures that
when hydrogens are disabled, the residue contains no hydrogens nor bonds
to hydrogens).

Add the protonation states to the list of possible Residues that can be
constructed.