Add Feature: Chain maintains Res in sorted order
This allows for residues to be added to the chain out of order, but
always returned or iterated over in sorted order by the residue IDs.
A use case can be when missing residues are added to a chain after it
build. Tests were added to TestPolypeptideChain to ensure this works.