Make Structure Classes Transformable
Add Feature: ability to write out to PDB. Initial implementation, writes
out the bare minimum so that the structure can be visualized with pymol.

Change Atom constructor to take doubles for the coordinates.

Change Sequence to iterate over Characters rather than Strings and add
constructor to Residue to take a character indicating which amino acid
to build.