Add Dihedral Angles
Dihedral angles can now be calculated for each Residue in
a PolypeptideChain. If a dihedral angle is undefined for a particular
residue, then 1000 is returned, otherwise the angle in degrees between
-180 and 180.

The version to use for jMath has been updated to 0.12.3 since a method
to calculate dihedral angles has been added to Vector3D in that version.

Add test in TestProtein to ensure that all dihedrals calculated for
1rop.pdb are correct. This is in TestProtein instead of
TestPolypeptideChain because a protein structure - 1rop - needs to
be constructed in order to test the dihedrals.