Add VirtualRibosome to package protein
This is an initial version of this class. It constructs a single chain
that is unfolded. When visualized, the chain is linear. All dihedral
angles are either 180 or -180. It doesn't check for clashing. It naively
moved each residue into place so that the bond angles are correct and
consecutive residues are on opposite sides of the chain.