Base GDT Scores on Num Residues in Reference
Base the Regions of Similarity scores on the number of residues in the
reference structure. Before this, it was based on the number of residues
in the alignment. This change grants more stability to the scores when
evaluating models produced for some native. They all use the same
baseline, the number of residues in that native, rather than different
baselines, the number of residues in the alignment between that model
and the native. Beyond stability, this also makes more sense. If a model
fails to represent residues that are in the native, it should be
penalized.