Add CMD Line Arg to JProtMetrics
Add the --chimera cmd line argument to specify that the program should
print out chimera scripts for the --ls or --gdt options. These scripts
can be used to color in chimera the regions of similarity found in the
structures being compared. The default behavior is to print out pymol
scripts.