Correct Defined Angle Triplets in Resource Files
The defined angle triplets were not consistent in the residues where
there is a branching. If there is a branch (eg, Leucine, Isoleucine, or
any residue with a ring), the angles of the branch should be defined and
A-B-C and A-B-D (rather than A-B-C and B-C-D). Both definitions are
valid, but consistency was needed, and given that valid angle triplets
are defined as three consecutive atoms in order from CA back, only the
former definition will yield correct manipulations with code
implementation.