Remove Protein Atom Graph, Streamline Method Name
Giving the responsibility of maintaining the Atom Graph to the Protein
added complication to the construction process and set up state
maintenance requirements that are difficult to maintain without setting
up an observer/observable pattern between a Protein and its components.
Any time a component of a Protein is modified, adding or removing
a PolypeptideChain, Residue, or Atom, the atoms graph would have to be
updated. It should rather be the responsibility of which ever class in
the future that needs an atom graph to build and maintain it.

Update the test to no longer test the graph since it is removed.

Also, change method containsChain() to contains().