Add manipulations Package
Externalize all the conformation manipulation code. Previously there
were redundant and copied parts in both PolypeptideChain and Residue.
The original thinking was that backbone manipulations would go in
PolypeptideChain and side chain manipulations would go in Residue, but
there is too much overlap. Modifying a Dihedral Angle is the same in
both cases, and the difference it was designed around (determining which
atoms to apply any Transformation to) could be abstracted away to the
same idea in both cases.

Backbone manipulations can affect every residue following the site of
the manipulation. Side chain manipulations can affect every atom in the
residue following the site of the manipulation. In both cases, the atoms
modified are those further from the N-Terminus than the site of the
manipulation. Within the residue, perform a depth first search from the
site back, and if the site is on the backbone, apply the transformation
to every residue after the current one as well.

The manipulators package includes the abstract super class
ConformationManipulator and two subclasses Cascade and Local
conformation manipulators. Cascade modifies all atoms after the site
of the manipulation. Local only modifies the atoms at the site of the
manipulation.

With the manipulators defined, VirtualRibosome was edited to make use of
them for the proper setting of dihedral and three atom angles after as
the Protein is built.

A tester was added for ConformationManipulator and the VirtualRibosome
tester was updated to match the expected values.