Make Sequence Abstract, Update Documentation
Make the align method in the subclasses of Sequence an abstract method
in the superclass and make the superclass abstract. Update the
documentation of SequenceAligner to make clear that clients only need to
explicitly use it if they want to specify custom gap penalty values for
the alignment algorithm.

Make Alignment constructor package private. Alignment should not be
constructed, only returned by SequenceAligner or Sequence::align.