Major Change to RoS-GDT Algorithm
This affects RoS-GDT, RoS-GDT-HA, and RoS-GDT-plot.

Its goal was to find regions of similarity where each subsequent region
completely contains the previous region. It was not accomplishing this
goal. The previous algorithm would run max clique on the neighborhood
containing the previous RoS (neighborhood: all nodes in the previous
RoS, all their neighbors, and all edges between those nodes), but that
is not stringent enough.

To make it stringent enough to ensure that the max clique found on it
completely contains the previous RoS, a neighbor is now only added to
the neighborhood if it has an edge to every node in the previous RoS.
Now the Max Clique found on that neighborhood is guaranteed to contain
the previous RoS, but we don't need to find the Max Clique on the entire
neighborhood. Just finding the max clique on the set of new nodes added
to the neighborhood will suffice. If we want to find the maximum set of
nodes to add to a RoS that maintains it as a clique, this set of nodes
is the max clique on those new neighborhood nodes, and since every node
in that set has an edge to every node in the old RoS, the union of the
two is still a clique.

Now the graphs that Max Clique is being solved on are smaller, and the
RoS-GDT algorithms are taking much less time.